+++ includes/minishell.h
@@ -42,6 +42,7 @@ typedef struct s_red
 }t_red;
 
 char	**modifsplit(char *input);
+char	**get_tokens(char *input);
 void	dollarfound_getlen(int *a, int *count, char *input);
 void	expand_var(char *orgn, char **end, int *a, int *b);
 int		ft_modstrcpy(char *orgn, char **end, int len);
@@ -50,14 +51,14 @@ void	ft_echo(char **list);
 //char	*adv_qm_rem(char *qm_str, int b_free);
 int		qm_error_detector(char *str);
 void	init_ms_env(char **env_vector, t_str **env_list);
-char	**env_list_to_vector(t_str **env_list);
-void	ft_lst_str_free(t_str **list);
-void	ft_lst_str_print(t_str *list);
-void	ft_lst_str_add_back(t_str **list, char *str);
-void	ft_lst_str_add_front(t_str **list, char *str);
-void	ft_lst_str_add_sorted(t_str **list, char *str);
-void	ft_lst_str_delete(t_str **list, char *str, size_t len);
-t_str	*ft_lst_str_get_str(t_str **list, const char *str);
+char	**lst_str_to_array(t_str **env_list);
+void	lst_str_free(t_str **list);
+void	lst_str_print(t_str *list);
+void	lst_str_add_back(t_str **list, char *str);
+void	lst_str_add_front(t_str **list, char *str);
+void	lst_str_add_sorted(t_str **list, char *str);
+void	lst_str_delete(t_str **list, char *str, size_t len);
+t_str	*lst_str_get_str(t_str **list, const char *str);
 void	exec_command(char **list, char **envp);
 char	*ft_strslashjoin(char const *s1, char const *s2);
 void	megafree(char ***list);
@@ -77,4 +78,5 @@ int		modstrcmp(char *str1, char *str2);		// ? Es como el strcmp, pero le llamo "
 char	*ft_getenv(t_str **env_list, const char *str);
 int		has_last_redirection_open(const char *str);
 int		has_pipe_redir_open(char **array);
+char	*ft_recursive_expand(char *str, t_str **env_list);
 #endif