+++ src/var_expansor.c
@@ -13,9 +13,9 @@
 #include "minishell.h"
 
 /*
- * This function recieves the adress of the first character of a 
- * variable's name ($varname), reads the variable name, gets 
- * it's value and returns it. 
+ * This function recieves the adress of the first character of a
+ * variable's name ($varname), reads the variable name, gets
+ * it's value and returns it.
  * Inputting the trimmed variable name is not required.
 */
 char	*getvarvalue(char *str)
@@ -40,11 +40,11 @@ char	*getvarvalue(char *str)
 }
 
 /*
- *This function recieves a string (orgn) and it's current position (a), 
+ *This function recieves a string (orgn) and it's current position (a),
  * and a pointer to a string (end) and it's current position.
- *The function is called once the parent function locates a variable name. 
- * This will read the variable name, get it's value, copy it into end and 
- * advance both indexes (a and b) to the last character read/written so 
+ *The function is called once the parent function locates a variable name.
+ * This will read the variable name, get it's value, copy it into end and
+ * advance both indexes (a and b) to the last character read/written so
  * the parent function can keep working.
 */
 void	expand_var(char *orgn, char **end, int *a, int *b)
@@ -65,8 +65,8 @@ void	expand_var(char *orgn, char **end, int *a, int *b)
 }
 
 /*
- * This function is similar to expand_var, but it will not copy the result 
- * of the variable into any string. It just counts the length of the 
+ * This function is similar to expand_var, but it will not copy the result
+ * of the variable into any string. It just counts the length of the
  * variable and advances the index to the string to the end of the var name.
 */
 void	dollarfound_getlen(int *a, int *count, char *input)
@@ -80,4 +80,86 @@ void	dollarfound_getlen(int *a, int *count, char *input)
 		*count += ft_strlen(value);
 	while ((ft_isalnum(input[*a]) || input[*a] == '_') && input[*a])
 		(*a)++;
+}
+
+/**
+ * * Get next index of a str to expand
+ * @param str		str to expand it content
+ * @return			str expanded
+*/
+static size_t get_start_expand(const char *str)
+{
+	int		quotes;
+	char	*start_expand;
+
+	quotes = NONE;
+	start_expand = (char *)str;
+	while (*start_expand != '\0')
+	{
+		if (quotes == NONE && *start_expand == '$')
+			return (start_expand - str);
+		else if (quotes == NONE && *start_expand == '\'')
+			quotes = SINGLE;
+		else if (quotes == SINGLE && *start_expand == '\'')
+			quotes = NONE;
+		start_expand++;
+	}
+	return (0);
+}
+
+
+/**
+ * * Replace environment variable of given str from start with some lenght
+ * @param str			str to expand it content
+ * @param env_list		environment list
+ * @param start_expand	index of start to replace
+ * @param len_expand	lenght to replace
+ * @return				str expanded
+*/
+static char *ft_expand_at(char *str, t_str **env_list, size_t start_expand, \
+							 size_t len_expand)
+{
+	char	*oldset;
+	char	*newset;
+	char	*out;
+
+	oldset = ft_substr(str, start_expand, len_expand);
+	newset = ft_getenv(env_list, &oldset[1]);
+	if (!newset)
+		out = ft_strreplaceat(str, oldset, "", start_expand);
+	else
+		out = ft_strreplaceat(str, oldset, newset, start_expand);
+	free(str);
+	free(oldset);
+	return (out);
+}
+
+/**
+ * * Given str expand environment variables ($ followed by characters) to their values
+ * * Input str should come malloc
+ * @param malloc_str	str to expand it content
+ * @param env_list		environment list
+ * @return				str expanded
+*/
+char	*ft_recursive_expand(char *malloc_str, t_str **env_list)
+{
+	size_t	start_expand;
+	size_t	len_expand;
+	char	*aux;
+
+	len_expand = 0;
+	start_expand = get_start_expand(malloc_str);
+	if (start_expand != 0)
+	{
+		len_expand = 1;
+		aux = &malloc_str[start_expand];
+		while (ft_isalnum(aux[len_expand]) || aux[len_expand] == '_')
+			len_expand++;
+		if (len_expand != 0)
+		{
+			malloc_str = ft_expand_at(malloc_str, env_list, start_expand, len_expand);
+			return (ft_recursive_expand(malloc_str, env_list));
+		}
+	}
+	return (malloc_str);
 }